Moves function to a moudle to be able to use it in another page
